refactoring

+++ src/ComplexityStatistics.php
@@ -9,16 +9,16 @@ use SebastianBergmann\Complexity\ComplexityCollection;
 final class ComplexityStatistics
 {
     /**
-     * @psalm-return array{minimum: non-negative-int, maximum: non-negative-int, average: float}
+     * @psalm-return array{minimum: int, maximum: int, average: float}
      */
     public static function from(ComplexityCollection $items): array
     {
         $values = collect($items)->map(fn ($item) => $item->cyclomaticComplexity());
 
         return [
-            'minimum' => $values->min() ?? 0,
-            'maximum' => $values->max() ?? 0,
-            'average' => $values->avg() ?? 0.0,
+            'minimum' => (int) ($values->min() ?? 0),
+            'maximum' => (int) ($values->max() ?? 0),
+            'average' => (float) ($values->avg() ?? 0.0),
         ];
     }
 }
